Separated power-down state from #active banks.
The power-down state of the memory was encoded in the variable mem_state. The same variable stored the number of active banks. I've separated the two now, since they are different, and for clarity should be tracked separately.

This exposed a bug in our libdrampower test. A memory controller may only enter a power-down mode when all commands are completed. This include auto-precharge commands. The test used to enter power-down immediately after a WRA command (before the write or auto-pre completed). The power-down command would set mem_state to a magic number. Once the PRE command was executed, it would change the mem_state (mem_state--). From that point the output was broken. I've updated the test by delaying the power-down appropriately.